+++ CHANGES.md
@@ -1,5 +1,29 @@
 # Version history and release notes
 
+## 3.3.8
+
+- Skip mappings for virtaul properties when mapping class or records.
+
+This is mainly for EF models compatibility. Example:
+
+```cs
+public class MyRefClass
+{
+    public int Value2 { get; set; }
+}
+
+public class MyClassWithVirtualRef
+{
+    public int Value1 { get; set; }
+    public virtual MyRefClass Ref { get; set; }
+}
+
+var result = connection.Read<MyClassWithVirtualRef>(@"select 1 as value1").First();
+Assert.Equal(1, result.Value1);
+Assert.Null(result.Ref);
+```
+
+
 ## 3.3.7
 
 - Added support anonymous class instance parameters: